Move mappings into common.py, add ids.py
