+++ .gitignore
@@ -5,4 +5,4 @@ compile_commands.json
 # Files that should be written from CAN spec
 can_validator/fsae_can_spec.h
 MY17_CAN_Library.c
-extra.c
+ids.h
