+++ can_validator/MY17_Can_Library.py
@@ -2,6 +2,7 @@ import sys
 sys.path.append("CAN_Api/src")
 from CANSpec import CANSpec
 from math import ceil
+import common
 
 spec_path = "CAN_Api/fsae_can_spec.yml"
 output_path = "../MY17_Can_Library.c"
@@ -9,43 +10,6 @@ base_path = "../MY17_Can_Library_BASE.txt"  # File with template code that's not
 special_cases = "special_cases.txt"  # Special cases of functions that aren't based off of CAN spec
 
 spec = CANSpec(spec_path)
-
-# Maps parts of segment names to their corresponding parts of names in structs
-field_name_mappings = {
-    "requested_torque": "torque",
-    "is_ok": "is_alive",
-    "wheel_speed": "wheel_speed_mRPM",
-    "right_throttle_pot": "accel_1_raw",
-    "left_throttle_pot": "accel_2_raw",
-    "front_brake_pressure": "brake_1_raw",
-    "rear_brake_pressure": "brake_2_raw",
-    "ave": "avg",
-    "min_cell_temp_id": "id_min_cell_temp",
-    "max_cell_temp_id": "id_max_cell_temp",
-    "min_cell_voltage_id": "id_min_cell_voltage",
-    "max_cell_voltage_id": "id_max_cell_voltage",
-    "error_type": "type",
-    "heartbeat_on":  "ok",
-    "request_type": "type",
-    "steering_pot": "steering_raw",
-    "soc_percentage": "soc",
-    "pack_energy": "energy_Wh"
-}
-
-# Maps parts of message names to their corresponding parts of names in the Can_MsgID_T enum
-msg_enum_name_mappings = {
-    "Front_Can_Node": "FrontCanNode",
-    "Rear_Can_Node": "RearCanNode",
-    "Raw_Values": "RawValues",
-    "Wheel_Speed": "WheelSpeed",
-    "Vcu_Bms_Heartbeat": "Vcu_BmsHeartbeat",
-    "Vcu_Dash_Heartbeat": "Vcu_DashHeartbeat",
-    "Cell_Temps": "CellTemps",
-    "Pack_Status": "PackStatus",
-    "Current_Sensor": "CurrentSensor",
-    "Driver_Output": "DriverOutput",
-    "Bms_Errors": "Bms_Error"
-}
 
 unused_messages = [
     "LV_BATTERY_VOLTAGE",
@@ -58,33 +22,6 @@ unused_messages = [
     "VCU_MC_MESSAGE",
     "MC_RESPONSE"
 ]
-
-
-def get_field_name(segment_name):
-    """
-    Find the name of the field in the message's struct for the particular segment.
-
-    :param segment_name: name of the segment
-    :return: name from struct
-    """
-    field_name = segment_name.lower()
-    for old, new in field_name_mappings.items():
-        field_name = field_name.replace(old, new)
-    return field_name
-
-
-def get_msg_enum_name(message_name):
-    """
-    Find the name of the message in the Can_MsgID_T enum for the particular CAN message
-    
-    :param message_name: message (from CAN spec)
-    :return: name used in the Can_MsgID_T enum
-    """
-    msg_enum_name = message_name.title()
-    for old, new in msg_enum_name_mappings.items():
-        msg_enum_name = msg_enum_name.replace(old, new)
-    return "Can_" + msg_enum_name
-
 
 with open(output_path, 'w') as f:
     # Copy over base
@@ -109,7 +46,7 @@ with open(output_path, 'w') as f:
             continue
         f.write(
             "    case " + message.name + "__id:\n" +
-            "      return " + get_msg_enum_name(message.name) + "_Msg;\n")
+            "      return " + common.get_msg_enum_name(message.name) + "_Msg;\n")
     f.write(
         "    default:\n" +
         "      return Can_Unknown_Msg;\n" +
@@ -122,7 +59,7 @@ with open(output_path, 'w') as f:
 
         # Write TO_CAN
         f.write(
-            "TO_CAN(" + get_msg_enum_name(message.name) + "){\n" +
+            "TO_CAN(" + common.get_msg_enum_name(message.name) + "){\n" +
             "  uint64_t bitstring = 0;\n")
 
         length = 0
@@ -160,11 +97,11 @@ with open(output_path, 'w') as f:
 
         # Write FROM_CAN
         f.write(
-            "FROM_CAN(" + get_msg_enum_name(message.name) + "){\n" +
+            "FROM_CAN(" + common.get_msg_enum_name(message.name) + "){\n" +
             "  uint64_t bitstring = 0;\n" +
             "  to_bitstring(can_in->data, &bitstring);\n")
         for segment_name, segment in message.segments.items():
-            field_name = get_field_name(segment_name)
+            field_name = common.get_field_name(segment_name)
             if message.name == "VCU_BMS_HEARTBEAT" and field_name == "state":
                 field_name = "alwaysTrue"
             if "CURRENT_SENSOR" in message.name:
